Update issue 3248
Fixing an issue in the degenerate case where we only have one possible flow direction but dinf requires a full facet to be defined.  Instead test for that case and set the direction appropriately for the facet index that we're pointing in.  It seems this issue was also occuring on the regular d_infinity algorithm so fixed it there too.


Former-commit-id: e0dc7cef35ef86ca10a9d0b049d6183383d5855a