Update issue 3248
Since the new flat resoultion algorithm doesn't modify the DEM we have to redefine what we mean by an outlet.  This is important for determing flow accumulation that works from the outlet up.  Numerous changes were added to the routing core most notably a find_outlets function that for the moment determines an outlet as a pixel that doens't have a flow direction defiend but does have a DEM.  We assume by this point the flow direction isn't defined because it can't drain into another pixel, thus turning into a point for a sink/outlet.


Former-commit-id: 8577284c8f7a461785b8640a8401a9e7920499ac