Fixes issue 3279
This was occurring because the user pixel size was larger than the AOI.  In part
this occurred because the user input an AOI that was unprojected but the pixel
size was listed as being in (meters) rather than the projected units of the
AOI polygon.  Updated the UI to indicate it's the units and added a more
helpful error message, AND allowed for floating point values to be passed in
for pixel sizes.


Former-commit-id: e465481b24c33d6a107152fd669cf93e1e059d76